refactor: move internal event code to event::internal
