+++ src/cursor/sys/unix.rs
@@ -4,7 +4,10 @@ use std::{
 };
 
 use crate::{
-    event::{filter::CursorPositionFilter, poll_internal, read_internal, InternalEvent},
+    event::{
+        filter::CursorPositionFilter,
+        internal::{poll_internal, read_internal, InternalEvent},
+    },
     terminal::{disable_raw_mode, enable_raw_mode, sys::is_raw_mode_enabled},
 };
