+++ src/event/read.rs
@@ -6,7 +6,9 @@ use crate::event::source::unix::UnixInternalEventSource;
 use crate::event::source::windows::WindowsEventSource;
 #[cfg(feature = "event-stream")]
 use crate::event::sys::Waker;
-use crate::event::{filter::Filter, source::EventSource, timeout::PollTimeout, InternalEvent};
+use crate::event::{
+    filter::Filter, internal::InternalEvent, source::EventSource, timeout::PollTimeout,
+};
 
 /// Can be used to read `InternalEvent`s.
 pub(crate) struct InternalEventReader {
