+++ src/event/stream.rs
@@ -14,8 +14,10 @@ use std::{
 use futures_core::stream::Stream;
 
 use crate::event::{
-    filter::EventFilter, lock_internal_event_reader, poll_internal, read_internal, sys::Waker,
-    Event, InternalEvent,
+    filter::EventFilter,
+    internal::{lock_internal_event_reader, poll_internal, read_internal, InternalEvent},
+    sys::Waker,
+    Event,
 };
 
 /// A stream of `Result<Event>`.
